requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.10.0
pyrebase4==4.7.1
firebase-admin==6.2.0
google-generativeai==0.8.2
Werkzeug==2.3.7
